refactor: extract single-invoice send pipeline to services/invoice_sending
Pure move of InvoiceSendConflict, _projected_items and
_send_single_invoice out of views/pre_billing.py so the upcoming
send-run worker and the HTTP views share one implementation
(batch-queue wave phase 1). View imports the service under the old
local names; behavior unchanged. Test patch targets repointed to the
service module. Views file 1115 -> 894 lines.
